requirements.txt
beautifulsoup4==4.9.3
pymongo==3.11.2
python-dotenv==0.15.0
python-telegram-bot==13.1
requests==2.25.1
